+++ app/Http/Controllers/DesaController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Desa;
+use App\Models\DetailDesa;
 use App\Models\Kecamatan;
+use App\Models\Tps;
 
 class DesaController extends Controller
 {
@@ -56,8 +58,20 @@ class DesaController extends Controller
 
     public function delete($id)
     {
-        $desa = Desa::find($id);
-        $desa->delete();
-        return redirect('/desa')->with('delete', 'Data Desa berhasil dihapus');
+        // cek apakah desa sudah terhubung detail desa
+
+        $detail_desa = DetailDesa::where('id_desa', $id)->first();
+        $cektps = Tps::where('id_desa', $id)->first();
+
+        if ($detail_desa) {
+            return redirect('/desa')->with('relasidetaildesa', 'Data Desa tidak bisa dihapus karena sudah terhubung dengan data TPS');
+        } elseif ($cektps) {
+            return redirect('/desa')->with('relasitps', 'Data Desa tidak bisa dihapus karena sudah terhubung dengan data TPS');
+        } else {
+
+            $desa = Desa::find($id);
+            $desa->delete();
+            return redirect('/desa')->with('delete', 'Data Desa berhasil dihapus');
+        }
     }
 }
